requirements.txt
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-Login==0.3.2
Flask-SQLAlchemy==2.1
Flask-Triangle==0.5.4
Flask-WTF==0.12
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
pbr==1.10.0
pkg-resources==0.0.0
rauth==0.7.2
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.15
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
Werkzeug==0.11.11
WTForms==2.1
